test(dsg): add id test
