+++ packages/amplication-data-service-generator/src/server/resource/dto/to-many/create-update-many-without-input/create-id.spec.ts
@@ -1,0 +1,11 @@
+import { createUpdateManyWithoutInputID } from ".";
+
+describe("Testing the id creation of the createUpdateMany", () => {
+  const pluralEntityName = "Posts";
+  const nestedEntityName = "Tag";
+  test("Testing the posts and tag example", () => {
+    expect(
+      createUpdateManyWithoutInputID(pluralEntityName, nestedEntityName).name
+    ).toBe(`${nestedEntityName}UpdateManyWithout${pluralEntityName}Input`);
+  });
+});
